restructure and weights initializations
