+++ model/config.py
@@ -1,13 +1,13 @@
 class TransformerConfig:
     def __init__(
         self,
-        input_vocab_size,
-        output_vocab_size,
-        d_seq,
-        d_embed,
-        n_heads,
-        n_layers,
-        d_ff,
+        input_vocab_size=100,
+        output_vocab_size=100,
+        d_seq=32,
+        d_embed=8,
+        n_heads=2,
+        n_layers=2,
+        d_ff=32 * 4,
         emb_dropout=0.1,
         ff_dropout=0.1,
         attention_dropout=0.1,
